Added Firebase upload for API call sequences and fixed an issue with the socket implementation
When calling the `/explore/random` endpoint, it now uploads the API call sequence along with SUT responses to firebase under the `test_api_calls` collection.

Fixed an issue with the `socket.io` implementation where it was adding a new listener on the `'connection'` event every time an API call was sent.