Implemented SUT restoration into the existing explore routes + bug fix and error message updates
The explore routes can now restore SUT state. On start up, if the call sequence exists in the DB, previous calls in the sequence will be rerun on the SUT. The previously run sequence will be remembered, and if the sequence name changes to another existing name the SUT state will be restored again.

`Note:` for this functionality to work correctly, the SUT needs to be reset whenever a schema is changed, or whenever the name of the API call sequence passed to the `/explore` route is changed. (Should not lead to any issues with the current visualizer setup though, just the DB).

- The application would crash if a schema name was not provided when setting/fetching - fixed
- When using the `/explore` route with the visualizer, there was a bug caused by the visualizer calling the parameters property `params` while the backend expected `paramters`. Now using the `/explore` route in the visualizer should pass the parameters as expected.
- Some error messages were updated in conjunction with Postman tests to be more informative of potential issues.

Co-Authored-By: Gabriel Kazai <144999882+GabrielKazai@users.noreply.github.com>