feat: #17 Endpoint for restoring SUT state (#18)
* Implemented SUT restoration into the existing explore routes + bug fix and error message updates

The explore routes can now restore SUT state. On start up, if the call sequence exists in the DB, previous calls in the sequence will be rerun on the SUT. The previously run sequence will be remembered, and if the sequence name changes to another existing name the SUT state will be restored again.

`Note:` for this functionality to work correctly, the SUT needs to be reset whenever a schema is changed, or whenever the name of the API call sequence passed to the `/explore` route is changed. (Should not lead to any issues with the current visualizer setup though, just the DB).

- The application would crash if a schema name was not provided when setting/fetching - fixed
- When using the `/explore` route with the visualizer, there was a bug caused by the visualizer calling the parameters property `params` while the backend expected `paramters`. Now using the `/explore` route in the visualizer should pass the parameters as expected.
- Some error messages were updated in conjunction with Postman tests to be more informative of potential issues.

Co-Authored-By: Gabriel Kazai <144999882+GabrielKazai@users.noreply.github.com>

* Improved check if previous calls has changed, attached warnings to the response

When rerunning a previous call sequence, the data object is also checked for potential changes.

If changes are detected during a rerun, a warning message will be attached to the final response (returning the call sequence for display on the timeline), saying which call was different (if status, what the status changed to).

* Implemented backend changes from client feedback

Automatic SUT restoration was removed. Sequences no longer maintain all call history from fresh SUT state, but instead just the calls from the previous exploration. When an additional exploration request is made with the same sequence name, the previous call sequence is overwritten on the database. The SUT state is no longer maintained automatically, so it will now be up to the user to decide whether to restart between explorations or not.

Between explorations with the same sequence name, a warnings array will be attached to the response body indicating what has changed (i.e., status, data, operation ID, call sequence length).

Co-Authored-By: Gabriel Kazai <144999882+GabrielKazai@users.noreply.github.com>

---------

Co-authored-by: Gabriel Kazai <144999882+GabrielKazai@users.noreply.github.com>