feat: #22 Routes for deleting existing schemas and sequences (#26)
* Added endpoints for deleting and renaming schemas and sequences

New routes:
- `PUT` apischema/rename/:schemaName/:newSchemaName
- `DELETE` apischema/delete/:schemaName
- `PUT` callsequence/rename/:sequenceName/:newSequenceName
- `DELETE` callsequence/delete/:sequenceName

Bug fix:
Spamming the `Submit schema` button in the visualizer with a new name would result in multiple duplicates due to async routes. Resolved by adding DB transactions for uploading new schemas (need to test more routes).

* Squashed commit of the following:

commit e0f98f27e048e5c29897c6a63f6f7f709eea974a
Author: Felix Bečeić <44879477+fbeceic@users.noreply.github.com>
Date:   Thu Dec 7 20:58:26 2023 +0100

    feat: #20 Favorites functionality

    Added an endpoint for toggling the favorite flag for a specific API schema call sequence.

    Added functions for retrieving a call sequence by name and updating a call sequence's favorite flag.

    ref: https://github.com/APIExplore/api-explore-backend/issues/20

* Explore bug fix

Fixed an issue with the explore route where if it fails to connect to the SUT it would delete the sequence prematurely resulting in an empty sequence.

* Added relationship analysis and sequence metrics

Changes are detailed in the readme under the #Exploration route.

* Resolved merge conflicts with main

* Fix for previous call sequence check

Ensures call sequences get checked when not running call by call mode. Reduces the amount of redundant length warnings when run this way. (will update later).